+++ bin/install-dev-env-scripts.sh
@@ -42,5 +42,6 @@ pip install "git+https://github.com/simonsdave/dev-env.git@$DEV_ENV_VERSION"
 
 download_script "${CLF_VERSION}" "run-all-spiders.sh"
 download_script "${CLF_VERSION}" "run-spider.sh"
+download_script "${CLF_VERSION}" "check-consistent-clf-version.sh"
 
 exit 0